\name{Pade-package}
\alias{Pade-package}
\docType{package}
\encoding{UTF-8}
\title{Padè Approximant Coefficients}
\description{Given a vector of Taylor series coefficients of sufficient length as input, the function returns the numerator and denominator coefficients for the Padè approximant of appropriate order.}
\author{Maintainer: Avraham Adler \email{Avraham.Adler@gmail.com}}
\keyword{package}
\keyword{NumericalMathematics}
