\name{Pade-package}
\alias{Pade-package}
\docType{package}
\title{
\packageTitle{Pade}
}
\description{
\packageDescription{Pade}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Pade}
\packageIndices{Pade}
}
\author{
\packageAuthor{Pade}

Maintainer: \packageMaintainer{Pade}
}
\keyword{ package }
\keyword{NumericalMathematics}
