\name{Pade-package}
\alias{Pade-package}
\docType{package}
\encoding{UTF-8}
\title{Padé Approximant Coefficients}
\description{Given a vector of Taylor series coefficients of sufficient length as input, the function returns the numerator and denominator coefficients for the Padé approximant of appropriate order (Baker, 1975) <ISBN:9780120748556>.}
\author{Maintainer: Avraham Adler \email{Avraham.Adler@gmail.com}}
\keyword{package}
\keyword{NumericalMathematics}
