% Copyright (c) 2015, Avraham Adler All rights reserved
% SPDX-License-Identifier: BSD-2-Clause OR GPL-2.0-or-later

\name{Pade-package}
\alias{Pade-package}
\docType{package}
\title{
\packageTitle{Pade}
}
\description{
\packageDescription{Pade}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Pade}
\packageIndices{Pade}
}
\author{
\packageAuthor{Pade}

Maintainer: \packageMaintainer{Pade}
}
\keyword{package}
\keyword{NumericalMathematics}
